Trait gstreamer_pbutils::prelude::DiscovererInfoExt

pub trait DiscovererInfoExt {
    fn copy(&self) -> DiscovererInfo;

    fn get_audio_streams(&self) -> Vec<DiscovererStreamInfo>;

    fn get_container_streams(&self) -> Vec<DiscovererStreamInfo>;

    fn get_duration(&self) -> ClockTime;

    fn get_live(&self) -> bool;

    fn get_misc(&self) -> Option<Structure>;

    fn get_missing_elements_installer_details(&self) -> Vec<String>;

    fn get_result(&self) -> DiscovererResult;

    fn get_seekable(&self) -> bool;

    fn get_stream_info(&self) -> Option<DiscovererStreamInfo>;

    fn get_stream_list(&self) -> Vec<DiscovererStreamInfo>;

    fn get_streams(&self, streamtype: Type) -> Vec<DiscovererStreamInfo>;

    fn get_subtitle_streams(&self) -> Vec<DiscovererStreamInfo>;

    fn get_tags(&self) -> Option<TagList>;

    fn get_toc(&self) -> Option<Toc>;

    fn get_uri(&self) -> Option<String>;

    fn get_video_streams(&self) -> Vec<DiscovererStreamInfo>;

    fn to_variant(&self, flags: DiscovererSerializeFlags) -> Option<Variant>;
}

Trait containing all DiscovererInfo methods.

Implementors

DiscovererInfo

Required Methods

fn copy(&self) -> DiscovererInfo

Returns

A copy of the DiscovererInfo

fn get_audio_streams(&self) -> Vec<DiscovererStreamInfo>

Finds all the DiscovererAudioInfo contained in self

Returns

A glib::List of matching DiscovererStreamInfo. The caller should free it with DiscovererStreamInfo::list_free.

fn get_container_streams(&self) -> Vec<DiscovererStreamInfo>

Finds all the DiscovererContainerInfo contained in self

Returns

A glib::List of matching DiscovererStreamInfo. The caller should free it with DiscovererStreamInfo::list_free.

fn get_duration(&self) -> ClockTime

Returns

the duration of the URI in gst::ClockTime (nanoseconds).

fn get_live(&self) -> bool

Feature: v1_14

Returns

whether the URI is live.

fn get_misc(&self) -> Option<Structure>

Deprecated

This functions is deprecated since version 1.4, use DiscovererInfoExt::get_missing_elements_installer_details

Returns

Miscellaneous information stored as a gst::Structure (for example: information about missing plugins). If you wish to use the gst::Structure after the life-time of self, you will need to copy it.

fn get_missing_elements_installer_details(&self) -> Vec<String>

Get the installer details for missing elements

Returns

An array of strings containing informations about how to install the various missing elements for self to be usable. If you wish to use the strings after the life-time of self, you will need to copy them.

fn get_result(&self) -> DiscovererResult

Returns

the result of the discovery as a DiscovererResult.

fn get_seekable(&self) -> bool

Returns

the whether the URI is seekable.

fn get_stream_info(&self) -> Option<DiscovererStreamInfo>

Returns

the structure (or topology) of the URI as a DiscovererStreamInfo. This structure can be traversed to see the original hierarchy. Unref with gst_discoverer_stream_info_unref after usage.

fn get_stream_list(&self) -> Vec<DiscovererStreamInfo>

Returns

the list of all streams contained in the info. Free after usage with DiscovererStreamInfo::list_free.

fn get_streams(&self, streamtype: Type) -> Vec<DiscovererStreamInfo>

Finds the DiscovererStreamInfo contained in self that match the given streamtype.

streamtype

a glib::Type derived from DiscovererStreamInfo

Returns

A glib::List of matching DiscovererStreamInfo. The caller should free it with DiscovererStreamInfo::list_free.

fn get_subtitle_streams(&self) -> Vec<DiscovererStreamInfo>

Finds all the DiscovererSubtitleInfo contained in self

Returns

A glib::List of matching DiscovererStreamInfo. The caller should free it with DiscovererStreamInfo::list_free.

fn get_tags(&self) -> Option<TagList>

Returns

all tags contained in the URI. If you wish to use the tags after the life-time of self, you will need to copy them.

fn get_toc(&self) -> Option<Toc>

Returns

TOC contained in the URI. If you wish to use the TOC after the life-time of self, you will need to copy it.

fn get_uri(&self) -> Option<String>

Returns

the URI to which this information corresponds to. Copy it if you wish to use it after the life-time of self.

fn get_video_streams(&self) -> Vec<DiscovererStreamInfo>

Finds all the DiscovererVideoInfo contained in self

Returns

A glib::List of matching DiscovererStreamInfo. The caller should free it with DiscovererStreamInfo::list_free.

fn to_variant(&self, flags: DiscovererSerializeFlags) -> Option<Variant>

Serializes self to a glib::Variant that can be parsed again through DiscovererInfo::from_variant.

Note that any gst::Toc (s) that might have been discovered will not be serialized for now.

flags

A combination of DiscovererSerializeFlags to specify what needs to be serialized.

Returns

A newly-allocated glib::Variant representing self.