Struct gstreamer_pbutils::Discoverer

pub struct Discoverer(_, _);

The Discoverer is a utility object which allows to get as much information as possible from one or many URIs.

It provides two APIs, allowing usage in blocking or non-blocking mode.

The blocking mode just requires calling Discoverer::discover_uri with the URI one wishes to discover.

The non-blocking mode requires a running glib::MainLoop iterating a glib::MainContext, where one connects to the various signals, appends the URIs to be processed (through Discoverer::discover_uri_async) and then asks for the discovery to begin (through Discoverer::start). By default this will use the GLib default main context unless you have set a custom context using glib::MainContext::push_thread_default.

All the information is returned in a DiscovererInfo structure.

Implements

glib::object::ObjectExt

Methods

impl Discoverer

pub fn new(timeout: ClockTime) -> Result<Discoverer, Error>

Creates a new Discoverer with the provided timeout.

timeout

timeout per file, in nanoseconds. Allowed are values between one second (GST_SECOND) and one hour (3600 * GST_SECOND)

Returns

The new Discoverer. If an error occurred when creating the discoverer, err will be set accordingly and None will be returned. If err is set, the caller must free it when no longer needed using glib::Error::free.

pub fn discover_uri(&self, uri: &str) -> Result<DiscovererInfo, Error>

pub fn discover_uri_async(&self, uri: &str) -> Result<(), BoolError>

pub fn start(&self)

pub fn stop(&self)

pub fn connect_discovered<F: Fn(&Discoverer, &DiscovererInfo, &Option<Error>) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_finished<F: Fn(&Discoverer) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_source_setup<F: Fn(&Discoverer, &Element) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_starting<F: Fn(&Discoverer) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Discoverer

pub fn set_property_timeout(&self, timeout: ClockTime)

pub fn get_property_timeout(&self) -> ClockTime

pub fn connect_property_timeout_notify<F: Fn(&Self) + Send + Sync + 'static>(
&self,
f: F
) -> SignalHandlerId